Update people.json
Thank you for putting up with my iterating on this. Trying to be accurate when starting a new job is hard.

Also, the Upstream Marketing Team is trying to help the SEO for the Kubernetes Contributor site. I intentionally omitted the noopener, noreffer tagging from kubernetes.dev. I hope that's okay as it is a community site.